Refactor and tidy things a bit
