+++ ckanext/dcat/controllers.py
@@ -1,0 +1,31 @@
+import json
+
+from ckan.plugins import toolkit
+
+if toolkit.check_ckan_version(min_version='2.1'):
+    BaseController = toolkit.BaseController
+else:
+    from ckan.lib.base import BaseController
+
+
+class DCATController(BaseController):
+
+    def dcat_json(self):
+
+        data_dict = {
+            'page': toolkit.request.params.get('page'),
+            'modified_since': toolkit.request.params.get('modified_since'),
+        }
+
+        try:
+            datasets = toolkit.get_action('dcat_datasets_list')({},
+                                                                data_dict)
+        except toolkit.ValidationError, e:
+            toolkit.abort(409, str(e))
+
+        content = json.dumps(datasets)
+
+        toolkit.response.headers['Content-Type'] = 'application/json'
+        toolkit.response.headers['Content-Length'] = len(content)
+
+        return content
